Rename client files to use TS extensions
