+++ client/src/pages/Auth/Register.test.tsx
