+++ client/src/pages/Categories.tsx
