+++ client/src/pages/Pagenotfound.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import Layout from "./../components/Layout";
+import Layout from "../components/Layout";
 
 const Pagenotfound = () => {
   return (
