+++ client/src/pages/Policy.tsx
@@ -1,4 +1,4 @@
-import Layout from "./../components/Layout";
+import Layout from "../components/Layout";
 
 const Policy = () => {
   return (
