+++ client/src/pages/Search.tsx
@@ -1,5 +1,5 @@
 import { useSearch } from "../context/search";
-import Layout from "./../components/Layout";
+import Layout from "../components/Layout";
 const Search = () => {
   const [values, setValues] = useSearch();
   return (
