+++ client/src/pages/admin/CreateProduct.tsx
@@ -2,14 +2,13 @@ import { Select } from "antd";
 import axios from "axios";
 import { useEffect, useState } from "react";
 import toast from "react-hot-toast";
-import { useNavigate, useParams } from "react-router-dom";
-import AdminMenu from "./../../components/AdminMenu";
-import Layout from "./../../components/Layout";
+import { useNavigate } from "react-router-dom";
+import AdminMenu from "../../components/AdminMenu";
+import Layout from "../../components/Layout";
 const { Option } = Select;
 
-const UpdateProduct = () => {
+const CreateProduct = () => {
   const navigate = useNavigate();
-  const params = useParams();
   const [categories, setCategories] = useState([]);
   const [name, setName] = useState("");
   const [description, setDescription] = useState("");
@@ -18,30 +17,7 @@ const UpdateProduct = () => {
   const [quantity, setQuantity] = useState("");
   const [shipping, setShipping] = useState("");
   const [photo, setPhoto] = useState("");
-  const [id, setId] = useState("");
 
-  //get single product
-  const getSingleProduct = async () => {
-    try {
-      const { data } = await axios.get(
-        `/api/v1/product/get-product/${params.slug}`
-      );
-      setName(data.product.name);
-      setId(data.product._id);
-      setDescription(data.product.description);
-      setPrice(data.product.price);
-      setPrice(data.product.price);
-      setQuantity(data.product.quantity);
-      setShipping(data.product.shipping);
-      setCategory(data.product.category._id);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-  useEffect(() => {
-    getSingleProduct();
-    //eslint-disable-next-line
-  }, []);
   //get all category
   const getAllCategory = async () => {
     try {
@@ -60,7 +36,7 @@ const UpdateProduct = () => {
   }, []);
 
   //create product function
-  const handleUpdate = async (e) => {
+  const handleCreate = async (e) => {
     e.preventDefault();
     try {
       const productData = new FormData();
@@ -68,16 +44,16 @@ const UpdateProduct = () => {
       productData.append("description", description);
       productData.append("price", price);
       productData.append("quantity", quantity);
-      photo && productData.append("photo", photo);
+      productData.append("photo", photo);
       productData.append("category", category);
-      const { data } = axios.put(
-        `/api/v1/product/update-product/${id}`,
+      const { data } = axios.post(
+        "/api/v1/product/create-product",
         productData
       );
       if (data?.success) {
         toast.error(data?.message);
       } else {
-        toast.success("Product Updated Successfully");
+        toast.success("Product Created Successfully");
         navigate("/dashboard/admin/products");
       }
     } catch (error) {
@@ -86,21 +62,6 @@ const UpdateProduct = () => {
     }
   };
 
-  //delete a product
-  const handleDelete = async () => {
-    try {
-      let answer = window.prompt("Are You Sure want to delete this product ? ");
-      if (!answer) return;
-      const { data } = await axios.delete(
-        `/api/v1/product/delete-product/${id}`
-      );
-      toast.success("Product DEleted Succfully");
-      navigate("/dashboard/admin/products");
-    } catch (error) {
-      console.log(error);
-      toast.error("Something went wrong");
-    }
-  };
   return (
     <Layout title={"Dashboard - Create Product"}>
       <div className="container-fluid m-3 p-3">
@@ -109,7 +70,7 @@ const UpdateProduct = () => {
             <AdminMenu />
           </div>
           <div className="col-md-9">
-            <h1>Update Product</h1>
+            <h1>Create Product</h1>
             <div className="m-1 w-75">
               <Select
                 bordered={false}
@@ -120,7 +81,6 @@ const UpdateProduct = () => {
                 onChange={(value) => {
                   setCategory(value);
                 }}
-                value={category}
               >
                 {categories?.map((c) => (
                   <Option key={c._id} value={c._id}>
@@ -141,19 +101,10 @@ const UpdateProduct = () => {
                 </label>
               </div>
               <div className="mb-3">
-                {photo ? (
+                {photo && (
                   <div className="text-center">
                     <img
                       src={URL.createObjectURL(photo)}
-                      alt="product_photo"
-                      height={"200px"}
-                      className="img img-responsive"
-                    />
-                  </div>
-                ) : (
-                  <div className="text-center">
-                    <img
-                      src={`/api/v1/product/product-photo/${id}`}
                       alt="product_photo"
                       height={"200px"}
                       className="img img-responsive"
@@ -208,20 +159,14 @@ const UpdateProduct = () => {
                   onChange={(value) => {
                     setShipping(value);
                   }}
-                  value={shipping ? "yes" : "No"}
                 >
                   <Option value="0">No</Option>
                   <Option value="1">Yes</Option>
                 </Select>
               </div>
               <div className="mb-3">
-                <button className="btn btn-primary" onClick={handleUpdate}>
-                  UPDATE PRODUCT
-                </button>
-              </div>
-              <div className="mb-3">
-                <button className="btn btn-danger" onClick={handleDelete}>
-                  DELETE PRODUCT
+                <button className="btn btn-primary" onClick={handleCreate}>
+                  CREATE PRODUCT
                 </button>
               </div>
             </div>
@@ -232,4 +177,4 @@ const UpdateProduct = () => {
   );
 };
 
-export default UpdateProduct;
+export default CreateProduct;
