+++ client/src/pages/admin/Products.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 import toast from "react-hot-toast";
 import { Link } from "react-router-dom";
 import AdminMenu from "../../components/AdminMenu";
-import Layout from "./../../components/Layout";
+import Layout from "../../components/Layout";
 const Products = () => {
   const [products, setProducts] = useState([]);
